Update server types and tsconfig
