--- server-types.d.ts
@@ -1,14 +1,0 @@
-import { NodeStatus } from '@kubernetes/client-node';
-
-declare global {
-  namespace Express {
-    interface Locals {
-      pods: any,
-      topPods: any,
-      topNodes: NodeStatus[],
-      nodeMetrics: any,
-      data: any,
-      result: any
-    }
-  }
-}
