+++ server/Controllers/kubeMetrics.ts
@@ -1,5 +1,7 @@
 require('dotenv').config();
 import { RequestHandler } from 'express';
+import { PodItem, PodStatusCount } from '../types/server-types'
+
 
 import {
   KubeConfig,
