--- server/server-types.ts
@@ -1,5 +1,0 @@
-export type ServerError = {
-  log: string;
-  status: number;
-  message: { err: string };
-};
