+++ server/tsconfig.json
@@ -12,8 +12,8 @@
     "noImplicitAny": true,
     "noEmit": true,
     "esModuleInterop": true,
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
@@ -24,23 +24,20 @@
       }
     ],
     "paths": {
-      "@/*": ["./*"]
+      "*": [
+        "../types/server-types.ts"
+      ],
+      "@/*": [
+        "./*"
+      ]
     }
   },
   "include": [
-    "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    ".next/types/**/*.ts",
-    "src/app/layout.jsx",
-    "src/app/app.jsx",
-    "src/app/dashboard/about/page.jsx",
-    "src/app/dashboard/alert/page.jsx",
-    "src/app/dashboard/pods/page.jsx",
-    "src/app/dashboard/report/page.jsx",
-    "src/app/dashboard/resource/page.jsx",
-    "src/app/dashboard/layout.tsx",
-    "src/app/ui/sidebar.tsx"
+    "./types/server-types.d.ts"
   ],
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
