+++ types/types.ts
@@ -1,0 +1,5 @@
+type PieChartData = { [key: string]: number };
+
+type PromMetricsData = any;
+type KubeMetricsstatus = any;
+type KubePodsStatus = any;
